+++ app/models/organisation.rb
@@ -21,12 +21,12 @@ class Organisation < ApplicationRecord
   validates :iati_reference,
     uniqueness: {case_sensitive: false},
     presence: true,
-    unless: proc { |organisation| organisation.matched_effort_provider? },
+    if: proc { |organisation| organisation.is_reporter? },
     format: {with: /\A[a-zA-Z]{2,}-[a-zA-Z]{3}-.+\z/, message: I18n.t("activerecord.errors.models.organisation.attributes.iati_reference.format")}
   validates :beis_organisation_reference, uniqueness: {case_sensitive: false}
   validates :beis_organisation_reference,
     presence: true,
-    unless: proc { |organisation| organisation.matched_effort_provider? },
+    if: proc { |organisation| organisation.is_reporter? },
     format: {with: /\A[A-Z]{2,5}\z/, message: I18n.t("activerecord.errors.models.organisation.attributes.beis_organisation_reference.format")}
 
   scope :sorted_by_name, -> { order(name: :asc) }